readers: structured readers and define reader type for all readers
