+++ opendrift/readers/basereader/__init__.py
@@ -18,10 +18,8 @@
 import sys
 import logging
 import copy
-from bisect import bisect_left
 from abc import abstractmethod, ABCMeta
 from datetime import datetime, timedelta
-from multiprocessing import Process, Manager, cpu_count
 
 from scipy.interpolate import LinearNDInterpolator
 import numpy as np
@@ -46,10 +44,6 @@ class BaseReader(Variables):
     return_block = True  # By default, all readers should be
                          # capable of returning blocks of data
 
-    # Default interpolation method, see function interpolate_block()
-    interpolation = 'linearNDFast'
-
-    convolve = None # Convolution kernel or kernel size
 
     verticalbuffer = 1  # To be overridden by application as needed
 
@@ -89,10 +83,6 @@ class BaseReader(Variables):
     def __init__(self):
         """Common constructor for all readers"""
         super().__init__()
-
-        # Dictionaries to store blocks of data for reuse (buffering)
-        self.var_block_before = {}  # Data for last timestep before present
-        self.var_block_after = {}   # Data for first timestep after present
 
         self.always_valid = False  # Set to True if a single field should
                                    # be valid at all times
@@ -147,12 +137,6 @@ class BaseReader(Variables):
                 self.spl_x((0,0)), self.spl_y((0,0))
 
         # Check if there are holes in time domain
-        if not hasattr(self, 'time_step'):
-            self.time_step = None
-        if not hasattr(self, 'start_time'):
-            self.start_time = None
-        if not hasattr(self, 'end_time'):
-            self.end_time = None
         if self.start_time is not None and self.time_step is not None:# and len(self.times) > 1:
             self.expected_time_steps = (
                 self.end_time - self.start_time).total_seconds() / (
@@ -179,7 +163,7 @@ class BaseReader(Variables):
         except:
             self.shape = None
 
-        self.set_buffer_size(max_speed=5)  # To be overriden by user/model
+        self.set_buffer_size(max_speed = 5.)
 
         # Check if there are east/north-oriented vectors
         for var in self.variables:
@@ -208,347 +192,10 @@ class BaseReader(Variables):
         else:
             return False
 
-    def set_buffer_size(self, max_speed, max_vertical_speed=None):
-        '''Adjust buffer to minimise data block size needed to cover elements'''
-        self.buffer = 0
-        pixelsize = self.pixel_size()
-        if pixelsize is not None:
-            if self.time_step is not None:
-                time_step_seconds = self.time_step.total_seconds()
-            else:
-                time_step_seconds = 3600  # 1 hour if not given
-            self.buffer = np.int(np.ceil(max_speed *
-                                         time_step_seconds /
-                                         pixelsize)) + 2
-            self.logger.debug('Setting buffer size %i for reader %s, assuming '
-                          'a maximum average speed of %g m/s.' %
-                          (self.buffer, self.name, max_speed))
-
-    def set_convolution_kernel(self, convolve):
-        """Set a convolution kernel or kernel size (of array of ones) used by `get_variables` on read variables."""
-        self.convolve = convolve
-
     def prepare(self, extent, start_time, end_time):
         """Prepare reader for given simulation coverage in time and space."""
         logging.debug('Nothing to prepare for ' + self.name)
         pass  # to be overriden by specific readers
-
-    def disabled_get_variables_interpolated(self, variables, profiles=None,
-                                   profiles_depth=None, time=None,
-                                   lon=None, lat=None, z=None,
-                                   block=False, rotate_to_proj=None):
-        """
-        `get_variables_interpolated` is the interface to the basemodel, and is
-        responsible for returning variables at the correct positions. This is done by:
-
-            1. Calling `__get_variables__` which,
-            2. calls `get_variables_derived`, which, finally
-            3. calls `get_variables`.
-
-        This function is responsible for setting up an interpolator,
-        `ReaderBlock` for regular gridded datasets.
-
-        Arguments:
-            variables: string, or list of strings (standard_name) of
-                requested variables. These must be provided by reader.
-
-            profiles: N/A
-
-            profiles_depth: N/A
-
-            time: datetime or None, time at which data are requested.
-                Can be None (default) if reader/variable has no time
-                dimension (e.g. climatology or landmask).
-
-            lon: N/A
-
-            lat: N/A
-
-            z: float or ndarray; vertical position (in meters, positive up)
-                of requested points.
-                default: 0 m (unless otherwise documented by reader)
-
-            block: bool, see return below
-
-            rotate_to_proj: N/A
-
-          Returns:
-            data: Dictionary
-                keywords: variables (string)
-                values:
-                    - 1D ndarray of len(x) if block=False. Nearest values
-                        (neighbour) of requested position are returned.
-                    - 3D ndarray encompassing all requested points in
-                        x,y,z domain if block=True. It is task of invoking
-                        application (OpenDriftSimulation) to perform
-                        interpolation in space and time.
-        """
-
-        self.timer_start('total')
-        self.timer_start('preparing')
-        # Raise error if time not not within coverage of reader
-        if not self.covers_time(time):
-            raise ValueError('%s is outside time coverage (%s - %s) of %s' %
-                             (time, self.start_time, self.end_time, self.name))
-
-        # Check which particles are covered (indep of time)
-        ind_covered, reader_x, reader_y = self.covers_positions(lon, lat, z)
-        if len(ind_covered) == 0:
-            raise ValueError(('All %s particles (%.2f-%.2fE, %.2f-%.2fN) ' +
-                              'are outside domain of %s (%s)') %
-                             (len(lon), lon.min(), lon.max(), lat.min(),
-                              lat.max(), self.name, self.coverage_string()))
-
-        # Find reader time_before/time_after
-        time_nearest, time_before, time_after, i1, i2, i3 = \
-            self.nearest_time(time)
-        self.logger.debug('Reader time:\n\t\t%s (before)\n\t\t%s (after)' %
-                      (time_before, time_after))
-        # For variables which are not time dependent, we do not care about time
-        static_variables = ['sea_floor_depth_below_sea_level', 'land_binary_mask']
-        if time == time_before or all(v in static_variables for v in variables):
-            time_after = None
-
-        if len(z) == 1 and len(lon) > 1:
-            z = z.copy()*np.ones(lon.shape)
-        z = z.copy()[ind_covered]  # Send values and not reference
-                                   # to avoid modifications
-        if block is False or self.return_block is False:
-            # Analytical reader, continous in space and time
-            self.timer_end('preparing')
-            env_before = self.get_variables_impl(variables, profiles,
-                                             profiles_depth,
-                                             time,
-                                             #time_before,
-                                             reader_x, reader_y, z,
-                                             block=block)
-            self.logger.debug('Fetched env-before')
-            self.timer_start('preparing')
-
-        else:
-            block_before = block_after = None
-            blockvariables_before = variables
-            blockvars_before = str(variables)
-            blockvariables_after = variables
-            blockvars_after = str(variables)
-            for blockvars in self.var_block_before:
-                if all(v in blockvars for v in variables):
-                    block_before = self.var_block_before[blockvars]
-                    blockvariables_before = block_before.data_dict.keys()
-                    blockvars_before = blockvars
-                    break
-                blockvariables_before = variables
-                blockvars_before = str(variables)
-            for blockvars in self.var_block_after:
-                if all(v in blockvars for v in variables):
-                    block_after = self.var_block_after[blockvars]
-                    blockvariables_after = block_after.data_dict.keys()
-                    blockvars_after = blockvars
-                    break
-
-            # Swap before- and after-blocks if matching times
-            if block_before is not None and block_after is not None:
-                if block_before.time != time_before:
-                    if block_after.time == time_before:
-                        block_before = block_after
-                        self.var_block_before[blockvars_before] = block_before
-                if block_after.time != time_after:
-                    if block_before.time == time_before:
-                        block_after = block_before
-                        self.var_block_after[blockvars_after] = block_after
-            # Fetch data, if no buffer is available
-            if block_before is None or \
-                    block_before.time != time_before:
-                self.timer_end('preparing')
-                reader_data_dict = \
-                    self.get_variables_impl(blockvariables_before, profiles,
-                                        profiles_depth, time_before,
-                                        reader_x, reader_y, z,
-                                        block=block)
-                self.timer_start('preparing')
-                self.var_block_before[blockvars_before] = \
-                    ReaderBlock(reader_data_dict,
-                                interpolation_horizontal=self.interpolation)
-                try:
-                    len_z = len(self.var_block_before[blockvars_before].z)
-                except:
-                    len_z = 1
-                self.logger.debug(('Fetched env-block (size %ix%ix%i) ' +
-                              'for time before (%s)') %
-                              (len(self.var_block_before[blockvars_before].x),
-                               len(self.var_block_before[blockvars_before].y),
-                               len_z, time_before))
-                block_before = self.var_block_before[blockvars_before]
-            if block_after is None or block_after.time != time_after:
-                if time_after is None:
-                    self.var_block_after[blockvars_after] = \
-                        block_before
-                else:
-                    self.timer_end('preparing')
-                    reader_data_dict = \
-                        self.get_variables_impl(blockvariables_after, profiles,
-                                            profiles_depth, time_after,
-                                            reader_x, reader_y, z,
-                                            block=block)
-                    self.timer_start('preparing')
-                    self.var_block_after[blockvars_after] = \
-                        ReaderBlock(
-                            reader_data_dict,
-                            interpolation_horizontal=self.interpolation)
-                    try:
-                        len_z = len(self.var_block_after[blockvars_after].z)
-                    except:
-                        len_z = 1
-
-                    self.logger.debug(('Fetched env-block (size %ix%ix%i) ' +
-                                  'for time after (%s)') %
-                                  (len(self.var_block_after[blockvars_after].x),
-                                   len(self.var_block_after[blockvars_after].y),
-                                   len_z, time_after))
-                    block_after = self.var_block_after[blockvars_after]
-
-            if (block_before is not None and block_before.covers_positions(
-                reader_x, reader_y) is False) or (\
-                block_after is not None and block_after.covers_positions(
-                    reader_x, reader_y) is False):
-                self.logger.warning('Data block from %s not large enough to '
-                                'cover element positions within timestep. '
-                                'Buffer size (%s) must be increased.' %
-                                (self.name, str(self.buffer)))
-
-            self.timer_end('preparing')
-            ############################################################
-            # Interpolate before/after blocks onto particles in space
-            ############################################################
-            self.timer_start('interpolation')
-            self.logger.debug('Interpolating before (%s) in space  (%s)' %
-                          (block_before.time, self.interpolation))
-            env_before, env_profiles_before = block_before.interpolate(
-                    reader_x, reader_y, z, variables,
-                    profiles, profiles_depth)
-
-            if (time_after is not None) and (time_before != time):
-                self.logger.debug('Interpolating after (%s) in space  (%s)' %
-                              (block_after.time, self.interpolation))
-                env_after, env_profiles_after = block_after.interpolate(
-                        reader_x, reader_y, z, variables,
-                        profiles, profiles_depth)
-
-            self.timer_end('interpolation')
-
-        #######################
-        # Time interpolation
-        #######################
-        self.timer_start('interpolation_time')
-        env_profiles = None
-        if (time_after is not None) and (time_before != time) and self.return_block is True:
-            weight_after = ((time - time_before).total_seconds() /
-                            (time_after - time_before).total_seconds())
-            self.logger.debug(('Interpolating before (%s, weight %.2f) and'
-                           '\n\t\t      after (%s, weight %.2f) in time') %
-                          (block_before.time, 1 - weight_after,
-                           block_after.time, weight_after))
-            env = {}
-            for var in variables:
-                # Weighting together, and masking invalid entries
-                env[var] = np.ma.masked_invalid((env_before[var] *
-                                                (1 - weight_after) +
-                                                env_after[var] * weight_after))
-            # Interpolating vertical profiles in time
-            if profiles is not None:
-                env_profiles = {}
-                self.logger.debug('Interpolating profiles in time')
-                # Truncating layers not present both before and after
-                numlayers = np.minimum(len(env_profiles_before['z']),
-                                       len(env_profiles_after['z']))
-                env_profiles['z'] = env_profiles_before['z'][0:numlayers+1]
-                for var in env_profiles_before.keys():
-                    if var == 'z':
-                        continue
-                    env_profiles_before[var]=np.atleast_2d(env_profiles_before[var])
-                    env_profiles_after[var]=np.atleast_2d(env_profiles_after[var])
-                    env_profiles[var] = (
-                        env_profiles_before[var][0:numlayers, :] *
-                        (1 - weight_after) +
-                        env_profiles_after[var][0:numlayers, :]*weight_after)
-            else:
-                env_profiles = None
-
-        else:
-            self.logger.debug('No time interpolation needed - right on time.')
-            env = env_before
-            if profiles is not None:
-                if 'env_profiles_before' in locals():
-                    env_profiles = env_profiles_before
-                else:
-                    # Copying data from environment to vertical profiles
-                    env_profiles = {'z': profiles_depth}
-                    for var in profiles:
-                        env_profiles[var] = np.ma.array([env[var], env[var]])
-        self.timer_end('interpolation_time')
-
-        ####################
-        # Rotate vectors
-        ####################
-        if rotate_to_proj is not None:
-            if self.simulation_SRS is True:
-                self.logger.debug('Reader SRS is the same as calculation SRS - '
-                              'rotation of vectors is not needed.')
-            else:
-                vector_pairs = []
-                for var in variables:
-                    for vector_pair in vector_pairs_xy:
-                        if var in vector_pair:
-                            counterpart = list(set(vector_pair) -
-                                               set([var]))[0]
-                            if counterpart in variables:
-                                vector_pairs.append(vector_pair)
-                            else:
-                                self.logger.warning('Missing component of vector pair, cannot rotate:' +
-                                         counterpart)
-                # Extract unique vector pairs
-                vector_pairs = [list(x) for x in set(tuple(x)
-                                for x in vector_pairs)]
-
-                if len(vector_pairs) > 0:
-                    self.timer_start('rotating vectors')
-                    for vector_pair in vector_pairs:
-                        env[vector_pair[0]], env[vector_pair[1]] = \
-                            self.rotate_vectors(reader_x, reader_y,
-                                                env[vector_pair[0]],
-                                                env[vector_pair[1]],
-                                                self.proj, rotate_to_proj)
-                        if profiles is not None and vector_pair[0] in profiles:
-                            env_profiles[vector_pair[0]], env_profiles[vector_pair[1]] = \
-                                    self.rotate_vectors (reader_x, reader_y,
-                                            env_profiles[vector_pair[0]],
-                                            env_profiles[vector_pair[1]],
-                                            self.proj,
-                                            rotate_to_proj)
-
-
-                    self.timer_end('rotating vectors')
-
-        # Masking non-covered pixels
-        self.timer_start('masking')
-        if len(ind_covered) != len(lon):
-            self.logger.debug('Masking %i elements outside coverage' %
-                          (len(lon)-len(ind_covered)))
-            for var in variables:
-                tmp = np.nan*np.ones(lon.shape)
-                tmp[ind_covered] = env[var].copy()
-                env[var] = np.ma.masked_invalid(tmp)
-                # Filling also fin missing columns
-                # for env_profiles outside coverage
-                if env_profiles is not None and var in env_profiles.keys():
-                    tmp = np.nan*np.ones((env_profiles[var].shape[0],
-                                          len(lon)))
-                    tmp[:, ind_covered] = env_profiles[var].copy()
-                    env_profiles[var] = np.ma.masked_invalid(tmp)
-
-        self.timer_end('masking')
-        self.timer_end('total')
-        return env, env_profiles
 
     def rotate_variable_dict(self, variables, proj_from='+proj=latlong', proj_to=None):
         for vectorpair in vector_pairs_xy:
@@ -560,49 +207,6 @@ class BaseReader(Variables):
                     variables['x'], variables['y'],
                     variables[vectorpair[0]], variables[vectorpair[1]],
                     proj_from, proj_to)
-
-    def nearest_time(self, time):
-        """Return nearest times before and after the requested time.
-
-        Returns:
-            nearest_time: datetime
-            time_before: datetime
-            time_after: datetime
-            indx_nearest: int
-            indx_before: int
-            indx_after: int
-        """
-        if self.start_time == self.end_time:
-            return self.start_time, self.start_time, None, 0, 0, 0
-        if self.start_time is None:
-            return None, None, None, None, None, None
-        if hasattr(self, 'times'):  # Time as array, possibly with holes
-            indx_before = np.max((0, bisect_left(self.times, time) - 1))
-            if self.times[indx_before + 1] == time:
-                # Correction needed when requested time exists in times
-                indx_before = indx_before + 1
-            time_before = self.times[indx_before]
-            indx_after = np.minimum(indx_before + 1,
-                                    len(self.times) - 1)  # At the end
-            time_after = self.times[indx_after]
-            if (time - time_before) < (time_after - time):
-                indx_nearest = indx_before
-            else:
-                indx_nearest = indx_after
-            nearest_time = self.times[indx_nearest]
-        else:  # Time step is constant (no holes)
-            if self.time_step is None:
-                return None, None, None, None, None, None
-            indx = float((time - self.start_time).total_seconds()) / \
-                float(self.time_step.total_seconds())
-            indx_nearest = int(round(indx))
-            nearest_time = self.start_time + indx_nearest*self.time_step
-            indx_before = int(np.floor(indx))
-            time_before = self.start_time + indx_before*self.time_step
-            indx_after = int(np.ceil(indx))
-            time_after = self.start_time + indx_after*self.time_step
-        return nearest_time, time_before, time_after,\
-            indx_nearest, indx_before, indx_after
 
     def index_of_closest_z(self, requested_z):
         """Return (internal) index of z closest to requested z.
@@ -804,7 +408,7 @@ class BaseReader(Variables):
             ry = np.float32(ry)
             rlon, rlat = self.xy2lonlat(rx, ry)
             data[variable] = np.ma.masked_invalid(data[variable])
-            if hasattr(self, 'convolve'):
+            if self.convole is not None:
                 from scipy import ndimage
                 N = self.convolve
                 if isinstance(N, (int, np.integer)):
