+++ opendrift/readers/reader_current_from_drifter.py
@@ -19,10 +19,10 @@ from datetime import datetime, timedelta
 import numpy as np
 import pyproj
 
-from opendrift.readers.basereader import BaseReader
+from opendrift.readers.basereader import BaseReader, StructuredReader
 
 
-class Reader(BaseReader):
+class Reader(BaseReader, StructuredReader):
 
     return_block = False
     variables = ['x_sea_water_velocity', 'y_sea_water_velocity']
@@ -70,7 +70,7 @@ class Reader(BaseReader):
                               (times[i+1] - times[i]).total_seconds())
         self.speed = np.array(self.speed)
         self.x_sea_water_velocity = self.speed*np.sin(np.radians(self.azimuth))
-        self.y_sea_water_velocity = self.speed*np.cos(np.radians(self.azimuth)) 
+        self.y_sea_water_velocity = self.speed*np.cos(np.radians(self.azimuth))
 
         # Subtract wind/stokes contribution
         if waves is not None:
@@ -111,7 +111,7 @@ class Reader(BaseReader):
 
     def get_variables(self, requested_variables, time=None,
                       x=None, y=None, z=None, block=False):
-        
+
         requested_variables, time, x, y, z, outside = self.check_arguments(
             requested_variables, time, x, y, z)
